requirements.txt
fastapi
flask
uvicorn
ffmpeg-python
python-multipart
docxtpl
python-docx
pydantic
torchaudio
dash
plotly
matplotlib
pandas
seaborn
torch
demucs
python-multipart
